News_Atlats_AI/requirements.txt
django==4.1.13
sqlparse==0.2.4
djongo==1.3.6
pymongo>=4.6.1
dnspython>=2.6.1
kafka-python>=2.0.2
feedparser>=6.0.10
python-dotenv>=1.0.0
Pillow>=10.2.0
requests>=2.31.0
beautifulsoup4>=4.12.2
nltk>=3.8.1
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0 
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-js-asset==2.0
django-elasticsearch-dsl
certifi